Allow several instances each handling a different stock.

+++ intraday.sh
@@ -1,31 +1,30 @@
 #!/bin/bash
 
-stock=183828
+stock=(183828 217041 183830 181870)
 
 # 0: buy-and-sell
 # 1: sell-and-buy
-mode=0
+mode=(0 0 0 0)
 
 #entering price
-price="88.50"
+price=("118.00" "81.50" "0" "0")
 
 # entering quantity
-quantity=730
+quantity=(730 830 1358 1054)
 
-trapped=0
+# The data to use for calibration
+calibration=(2012-01-30 2012-01-30 2012-01-30 2012-01-30)
+
+status=(0 0 0 0)
 
 wd=/home/xxie/work/avanza/data_extract/intraday
 
-function start_trading {
-    if [ "$trapped" == "0" ]; then
-	./intraday -s $stock -m $mode -q $quantity -p $price
-    else
-	./intraday -s $stock -m $mode -q $quantity -p $price -t
-    fi
-}
+# function start_trading {
+#     ./intraday -s $1 -m $2 -t $3 -q $4 -p $5 -c $6 -t $7
+# }
 
 function stop_trading {
-    pid=`ps -C intraday -o pid=,cmd= | grep $stock | grep -o '^ *[0-9]\+'`
+    pid=`ps -C intraday -o pid=,cmd= | grep $1 | grep -o '^ *[0-9]\+'`
     if [ -z "$pid" ]; then
 	echo "None is running."
 	return 0
@@ -40,7 +39,7 @@ function stop_trading {
 
 if [ "$wd" != `pwd` ]; then cd $wd; fi
 
-for dir in transactions records logs probdist; do
+for dir in transactions records logs; do
     if [ ! -d $dir ]; then
 	mkdir $dir
     fi
@@ -48,39 +47,60 @@ done
 
 if [ -z "$1" ]; then
     echo "Usage: $0 COMMAND" 1>&2
-    exit 1
+    exit -1
 fi
 
 # start_trading
 command="$1"
-shift
 
 case $command in
 start)
-	if ps -C intraday -o cmd= | grep -q $stock; then
-	    echo "Already running."
-	    exit 0
+	if [ -n "$2" ]; then
+	    for ((i=0; i<${#stock[@]}; i++)); do
+		if [ "${stock[$i]}" != "$2" ]; then
+		    continue;
+		fi
+		./intraday -s ${stock[$i]} -m ${mode[$i]} -t ${status[$i]} \
+		    -q ${quantity[$i]} -p ${price[$i]} -c ${calibration[$i]} &
+		break;
+	    done
+	    exit 0;
 	fi
+	
+	for ((i=0; i<${#stock[@]}; i++)); do
+	    if ps -C intraday -o cmd= | grep -q ${stock[$i]}; then
+		echo "${stock[$i]} Already running."
+		continue;
+	    else
+		./intraday -s ${stock[$i]} -m ${mode[$i]} -t ${status[$i]} \
+		    -q ${quantity[$i]} -p ${price[$i]} -c ${calibration[$i]} &
+	    fi
+	done
 	if [ `pgrep -c alarm.sh` -eq 0 ]; then
 	    ./alarm.sh &
 	fi
-	start_trading
 	;;
 stop)
-	stop_trading
+	if [ -n "$2" ]; then
+	    for ((i=0; i<${#stock[@]}; i++)); do
+		if [ "${stock[$i]}" != "$2" ]; then
+		    continue;
+		fi
+		stop_trading ${stock[$i]}
+		break;
+	    done
+	    exit 0;
+	fi
+	
+	for ((i=0; i<${#stock[@]}; i++)); do
+	    stop_trading ${stock[$i]}
+	done
 	if [ `pgrep -c intraday` -eq 0 ]; then
 	    killall alarm.sh &> /tmp/null
 	fi
 	;;
-restart)
-	stop_trading
-	if [ `pgrep -c alarm.sh` -eq 0 ]; then
-	    ./alarm.sh &
-	fi
-	start_trading
-	;;
 *)
 	echo "Unknown command \"$command\""
-	exit 1
+	exit -2
 esac
